model refactoring
